Move CloseView to its own file
